backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
pymongo==4.6.1
google-api-python-client==2.79.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests==2.28.2
python-dotenv==1.0.0
pydantic==2.10.6
